+++ app/models/game_day_secretary_link_game_day.rb
@@ -1,0 +1,4 @@
+class GameDaySecretaryLinkGameDay < ApplicationRecord
+  belongs_to :game_day_secretary_link
+  belongs_to :game_day
+end